django_server/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-environ==0.11.2
requests==2.32.3
sqlparse==0.5.1
APScheduler~=3.10.4
colorlog~=6.8.2
django-oauth-toolkit~=3.0.1
daphne~=4.1.2
python-dateutil~=2.9.0.post0
paho-mqtt~=2.1.0
aiohttp~=3.12.13
cryptography~=44.0.3
aiofiles~=24.1.0
Pillow~=11.0.0
PyP100~=0.1.2